Create HTU_base_lattice_matched.lte
adding the base_matched lattice.

usage note: the HTU_match_fodo.ele file can be modified to use the base_lattice.lte and you can allow it to vary additional parameters like individual pmq separations. The optimal match conditions will of course depend on the e-beam source parameters which are not known a priori in experiment. So, the base_lattice.lte is meant to represent the physical system as it stands to study beam tracking and similar things under present conditions.