adding examles, and updated poetry.toml to use active master version of rsopt and functioning version of sirepo
Had some conflicts between rsopt and sirepo when using current versions for install. Now using current version of rsopt which includes bug fixes we identified, and a version of sirepo that was found compatible. Note, i thought it was just one commit behind, but it's more like ~20.

Added functional example for rsopt.