Refactor QuadScanProgram, generalizing output of measure()
Thinking that it's not necessarily Twiss parameters that is output for
every scan program - think full images for phase space reconstruction -
this commit reverts the method name to measure() and adds a few types.

For GEECSPythonAPIQuadScan, this specifies that this program is for
Twiss parameters by renaming to GEECSPythonAPITwissQuadScan, and having
measure() output type GEECSPythonAPITwissQuadScanMeasurement.